Move syllable-related helper files to language processing helpers and adjust incorrect file paths
