+++ packages/yoastseo/src/languageProcessing/helpers/syllables/syllableCountIterator.js
