+++ packages/yoastseo/src/languageProcessing/helpers/syllables/syllableCountStep.js
@@ -1,6 +1,6 @@
 import { isUndefined } from "lodash-es";
 
-import arrayToRegex from "../../../researches/stringProcessing/createRegexFromArray.js";
+import arrayToRegex from "../regex/createRegexFromArray.js";
 
 /**
  * Constructs a language syllable regex that contains a regex for matching syllable exclusion.
