+++ packages/yoastseo/src/scoring/assessments/readability/paragraphTooLongAssessment.js
@@ -4,7 +4,7 @@ import { inRangeEndInclusive as inRange } from "../../../helpers/inRange";
 import isParagraphTooLong from "../../../helpers/isValueTooLong";
 import marker from "../../../markers/addMark";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
-import { stripBlockTagsAtStartEnd as stripHTMLTags } from "../../../../researches/stringProcessing/stripHTMLTags";
+import { stripBlockTagsAtStartEnd as stripHTMLTags } from "../../../../../helpers/src/strings/stripHTMLTags";
 import AssessmentResult from "../../../values/AssessmentResult";
 import Mark from "../../../values/Mark";
