+++ packages/yoastseo/src/scoring/assessments/readability/passiveVoiceAssessment.js
@@ -5,7 +5,7 @@ import getLanguageAvailability from "../../../helpers/_todo/getLanguageAvailabil
 import { inRangeEndInclusive as inRange } from "../../../helpers/inRange";
 import marker from "../../../markers/addMark";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
-import { stripIncompleteTags as stripTags } from "../../../../researches/stringProcessing/stripHTMLTags";
+import { stripIncompleteTags as stripTags } from "../../../../../helpers/src/strings/stripHTMLTags";
 import AssessmentResult from "../../../values/AssessmentResult";
 import Mark from "../../../values/Mark";
