+++ packages/yoastseo/src/scoring/assessments/readability/sentenceLengthInTextAssessment.js
@@ -6,7 +6,7 @@ import formatNumber from "../../../helpers/formatNumber";
 import { inRangeEndInclusive as inRange } from "../../../helpers/inRange";
 import addMark from "../../../markers/addMark";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
-import { stripIncompleteTags as stripTags } from "../../../../researches/stringProcessing/stripHTMLTags";
+import { stripIncompleteTags as stripTags } from "../../../../../helpers/src/strings/stripHTMLTags";
 import AssessmentResult from "../../../values/AssessmentResult";
 import Mark from "../../../values/Mark";
