+++ packages/yoastseo/src/scoring/assessments/readability/wordComplexityAssessment.js
@@ -3,7 +3,7 @@ import { filter, flatMap, flatten, forEach, zip } from "lodash-es";
 import formatNumber from "../../../helpers/formatNumber";
 import addMark from "../../../markers/addMark";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
-import removeSentenceTerminators from "../../../../researches/stringProcessing/removeSentenceTerminators";
+import removeSentenceTerminators from "../../../languageProcessing/helpers/sanitize/removeSentenceTerminators";
 import AssessmentResult from "../../../values/AssessmentResult";
 import Mark from "../../../values/Mark";
